+++ gfx/drivers_font/d3d10_font.c
@@ -85,9 +85,9 @@ static void d3d10_font_free(void* data, bool is_threaded)
    free(font);
 }
 
-static int d3d10_font_get_message_width(void* data, const char* msg, unsigned msg_len, float scale)
+static int d3d10_font_get_message_width(void* data, const char* msg, size_t msg_len, float scale)
 {
-   unsigned i;
+   int i;
    int      delta_x                 = 0;
    const struct font_glyph* glyph_q = NULL;
    d3d10_font_t* font               = (d3d10_font_t*)data;
@@ -122,7 +122,7 @@ static void d3d10_font_render_line(
       d3d10_video_t *d3d10,
       d3d10_font_t*       font,
       const char*         msg,
-      unsigned            msg_len,
+      size_t              msg_len,
       float               scale,
       const unsigned int  color,
       float               pos_x,
@@ -250,7 +250,7 @@ static void d3d10_font_render_message(
    if (!font->font_driver->get_line_metrics ||
        !font->font_driver->get_line_metrics(font->font_data, &line_metrics))
    {
-      unsigned msg_len = strlen(msg);
+      size_t msg_len = strlen(msg);
       if (msg_len <= (unsigned)d3d10->sprites.capacity)
          d3d10_font_render_line(d3d10,
                font, msg, msg_len, scale, color, pos_x, pos_y,
@@ -263,8 +263,8 @@ static void d3d10_font_render_message(
    for (;;)
    {
       const char* delim = strchr(msg, '\n');
-      unsigned msg_len  = delim ?
-         (unsigned)(delim - msg) : strlen(msg);
+      size_t msg_len  = delim ?
+         (delim - msg) : strlen(msg);
 
       /* Draw the line */
       if (msg_len <= (unsigned)d3d10->sprites.capacity)
